+++ tests/misc/programs/misc/src/lib.rs
@@ -152,4 +152,16 @@ pub mod misc {
         assert!(ctx.accounts.token.mint == ctx.accounts.mint.key());
         Ok(())
     }
+
+    pub fn test_composite_payer(ctx: Context<TestCompositePayer>) -> ProgramResult {
+        ctx.accounts.composite.data.data = 1;
+        ctx.accounts.data.udata = 2;
+        ctx.accounts.data.idata = 3;
+        Ok(())
+    }
+
+    pub fn test_init_associated_token(ctx: Context<TestInitAssociatedToken>) -> ProgramResult {
+        assert!(ctx.accounts.token.mint == ctx.accounts.mint.key());
+        Ok(())
+    }
 }